requirements.txt
moviepy==2.1.1
streamlit==1.40.2
edge_tts==6.1.19
fastapi==0.115.6
uvicorn==0.32.1
openai==1.56.1
faster-whisper==1.1.0
loguru==0.7.2
google.generativeai==0.8.3
dashscope==1.20.14
g4f==0.3.8.1
azure-cognitiveservices-speech==1.41.1
redis==5.2.0
python-multipart==0.0.19
streamlit-authenticator==0.4.1
pyyaml
google-api-python-client==2.118.0
google-auth-oauthlib==1.2.0
